// src/models/PlanetDetails.tsx
export interface PlanetDetailsModel {
    climate: string;
    created: string;
    diameter: number;
    edited: string;
    films: string[];
    gravity: number;
    name: string;
    orbital_period: number;
    population: number;
    residents: number;
    rotation_period: number;
    surface_water: number;
    terrain: string;
    url: string;
}